Fixes in the getPartitioningAncestors Server functionality.
 - Made some improvements in the `SeqImplicits` scala classes.
 - Made it much more streamlined and simple.
 - Added a new file in the model called PartitioningResult.scala containing `PartitioningResult`
