+++ server/src/main/scala/za/co/absa/atum/server/model/PartitioningResult.scala
@@ -1,0 +1,42 @@
+/*
+ * Copyright 2021 ABSA Group Limited
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package za.co.absa.atum.server.model
+
+import io.circe.{Decoder, DecodingFailure, Json}
+import za.co.absa.atum.model.dto.{PartitionDTO, PartitioningWithIdDTO}
+import za.co.absa.atum.server.implicits.SeqImplicits.SeqEnhancements
+
+abstract class PartitioningResult(id: Long, partitioning: Json, author: String) {
+
+  protected def toPartitioningWithIdDTO(implicit
+                                        decoder: Decoder[PartitioningForDB]
+                                       ): Either[DecodingFailure, PartitioningWithIdDTO] = {
+    decoder.decodeJson(partitioning).map { partitioningForDB =>
+      val partitioningDTO = partitioningForDB.keys.map { key =>
+        PartitionDTO(key, partitioningForDB.keysToValuesMap(key))
+      }
+      PartitioningWithIdDTO(id, partitioningDTO, author)
+    }
+  }
+}
+
+object PartitioningResult {
+
+  def resultsToPartitioningWithIdDTOs(results: Seq[PartitioningResult]): Either[DecodingFailure, Seq[PartitioningWithIdDTO]] = {
+    results.seqDecode(_.toPartitioningWithIdDTO)
+  }
+}
